Update README.md
I added a section in the Project Overview to match the previous one talking about what we did in Milestone 1 for milestone 2. If everyone could add on bullet point of what they did that would be great. I also Added what I did for Milestone 2 into the table and made a table forecasting what we will do for Milestone 3 (something we were missing last time).